refactor(chatgpt): Move `createEvaluation` method from `ChatGptEvaluationService` to `ChatGptCompletionService`
