+++ server/src/main/kotlin/org/elaastic/ai/evaluation/chatgpt/api/DefaultChatGptCompletionService.kt
@@ -19,19 +19,33 @@ package org.elaastic.ai.evaluation.chatgpt.api
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.module.SimpleModule
+import org.elaastic.activity.response.Response
+import org.elaastic.ai.evaluation.chatgpt.ChatGptEvaluation
+import org.elaastic.ai.evaluation.chatgpt.ChatGptEvaluationData
+import org.elaastic.ai.evaluation.chatgpt.ChatGptEvaluationRepository
 import org.elaastic.ai.evaluation.chatgpt.ChatGptEvaluationService
+import org.elaastic.ai.evaluation.chatgpt.ChatGptEvaluationStatus
+import org.elaastic.ai.evaluation.chatgpt.PromptData
+import org.elaastic.ai.evaluation.chatgpt.prompt.ChatGptPrompt
+import org.elaastic.ai.evaluation.chatgpt.prompt.ChatGptPromptService
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Profile
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.MediaType
+import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Propagation
+import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.client.RestTemplate
+import java.util.logging.Level
 import java.util.logging.Logger
 
 @Profile("chatgpt")
 @Service
-class DefaultChatGptCompletionService(
+open class DefaultChatGptCompletionService(
+    val chatGptPromptService: ChatGptPromptService,
+    val chatGptEvaluationRepository: ChatGptEvaluationRepository,
     val restTemplate: RestTemplate,
     val objectMapper: ObjectMapper,
     @Value("\${chatgptapi.token}")
@@ -48,6 +62,61 @@ class DefaultChatGptCompletionService(
         }
         // logger for this class
         val logger = Logger.getLogger(ChatGptEvaluationService::class.java.name)
+    }
+
+    /**
+     * Create a ChatGPT evaluation for a response. The evaluation is created asynchronously.
+     *
+     * @param response the response to evaluate
+     * @param language the language of the evaluation
+     * @param chatGptExistingEvaluation the existing evaluation if it exists
+     * @return the created evaluation
+     */
+    @Async
+    @Transactional(propagation = Propagation.NEVER)
+    override fun createEvaluation(
+        response: Response,
+        language: String,
+        chatGptExistingEvaluation: ChatGptEvaluation?
+    ): ChatGptEvaluation {
+        // get the default prompt for the language
+        val chatGptDefaultPrompt = chatGptPromptService.getPrompt(language)
+        // Initialization of the evaluation
+        val chatGptEvaluation = chatGptExistingEvaluation ?: ChatGptEvaluation(response = response)
+        chatGptEvaluation.prompt = chatGptDefaultPrompt
+        markEvaluationAsPending(chatGptEvaluation)
+        // build the prompt
+        val prompt = buildThePrompt(chatGptDefaultPrompt, response)
+        // build the evaluation
+        try {
+            // get the response from ChatGPT
+            logger.info("Generating response with ChatGPT for response ${response.id}")
+            logger.fine("Prompt: $prompt")
+            val generatedResponse = getChatGptResponse(
+                listOf(ChatGptApiMessageData(role = "user", content = prompt)),
+            ).messageList.first().content
+            logger.info("Response generated with ChatGPT for response ${response.id}")
+            logger.fine("Generated response: $generatedResponse")
+            // convert the generated response to a ChatGptEvaluationData object
+            val chatGptEvaluationData = ObjectMapper().readValue(
+                generatedResponse,
+                ChatGptEvaluationData::class.java
+            )
+            // finalize the evaluation
+            chatGptEvaluation.status = ChatGptEvaluationStatus.DONE.name
+            chatGptEvaluation.grade = chatGptEvaluationData.grade
+            chatGptEvaluation.annotation = chatGptEvaluationData.annotation
+        } catch (e: Exception) {
+            chatGptEvaluation.status = ChatGptEvaluationStatus.ERROR.name
+            logger.log(Level.SEVERE, "Error while evaluating response with ChatGPT: ${e.message}", e)
+        }
+        return chatGptEvaluationRepository.save(chatGptEvaluation)
+    }
+
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
+    open fun markEvaluationAsPending(chatGptEvaluation: ChatGptEvaluation): ChatGptEvaluation {
+        chatGptEvaluation.status = ChatGptEvaluationStatus.PENDING.name
+        return chatGptEvaluationRepository.saveAndFlush(chatGptEvaluation)
     }
 
     /**
@@ -78,4 +147,31 @@ class DefaultChatGptCompletionService(
         return objectMapper.writeValueAsString(requestBody)
     }
 
+    private fun buildThePrompt(
+        chatGptPrompt: ChatGptPrompt,
+        response: Response
+    ): String {
+        val questionTitle = response.statement.title
+        val questionStatement = response.statement.content
+        val teacherExplanation = response.statement.expectedExplanation
+        val studentExplanation = response.explanation
+
+        requireNotNull(teacherExplanation) { throw IllegalArgumentException("Error: You must define an expected explanation to create a ChatGPT evaluation") }
+        requireNotNull(studentExplanation) { throw IllegalArgumentException("Error: No explanation to evaluate") }
+
+        // add to the prompt the title, the question content, the teacher explanation and the student explanation as a json object
+        val promptData = PromptData(
+            questionTitle = questionTitle,
+            questionStatement = questionStatement,
+            teacherExplanation = teacherExplanation,
+            studentExplanation = studentExplanation,
+            studentChoices = response.learnerChoice,
+            studentScoreBasedOnChoices = response.score,
+        )
+
+        val objectMapper = ObjectMapper()
+        val jsonObject = objectMapper.writeValueAsString(promptData)
+
+        return chatGptPrompt.content + "\n" + jsonObject
+    }
 }
